requirements.txt
django
django-cors-headers
django-filter
djangorestframework
coreapi
